+++ src/components/UI/Logo.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const Logo = () => (
+  <div>
+    <p className="logo">ShopMe</p>
+  </div>
+);
+
+export default Logo;
