+++ src/components/search/Search.css
@@ -1,0 +1,16 @@
+@import url('https://fonts.googleapis.com/css?family=Pacifico');
+* {
+  box-sizing: border-box;
+  margin: 0;
+  padding: 0;
+}
+
+body {
+  width: 1180px;
+  margin: 0 auto;
+}
+
+.logo {
+  font-family: 'Pacifico', cursive;
+  font-size: 2rem;
+}
