+++ src/core/Routes.jsx
@@ -1,16 +1,16 @@
 import React from 'react';
-import {BrowserRouter, Route, Switch} from 'react-router-dom';
+import { BrowserRouter, Route, Switch } from 'react-router-dom';
 
 import ScreensExamplesHello from 'screens/Examples/Hello';
 import ScreensExamplesFetch from 'screens/Examples/Fetch';
-import Search from 'components/search/search-component';
+import ScreensSearch from 'screens/Search/Search';
 
 export default() => (
   <BrowserRouter>
     <Switch>
-      <Route exact path="/" component={Search}/>
-      <Route path="/examples/hello" component={ScreensExamplesHello}/>
-      <Route path="/examples/fetch" component={ScreensExamplesFetch}/>
+      <Route exact path="/" component={ScreensSearch} />
+      <Route path="/examples/hello" component={ScreensExamplesHello} />
+      <Route path="/examples/fetch" component={ScreensExamplesFetch} />
     </Switch>
   </BrowserRouter>
 );
